requirements.txt
streamlit
plotly
pandas
numpy
scikit-learn
matplotlib
seaborn
pillow
tqdm
tensorflow
